+++ test/testrunner.html
@@ -24,7 +24,11 @@ mocha.reporter('html');
 </script>
 
 <!-- Tests -->
-<script src="test.js"></script>
+<script src="member/normaliseAria.js"></script>
+<script src="instance/identify.js"></script>
+<script src="instance/role.js"></script>
+<script src="instance/addRole.js"></script>
+<script src="instance/removeRole.js"></script>
 
 <script>
 // Only tests run in real browser, injected script run if options.run == true
